collection: extract AND/OR collection function's same code to trait
