+++ src/Collection/Functions/DisjunctionOperatorFunction.php
@@ -15,6 +15,9 @@ use Nextras\Orm\Entity\IEntity;
 
 class DisjunctionOperatorFunction implements IArrayFunction, IQueryBuilderFunction
 {
+	use JunctionFunctionTrait;
+
+
 	/** @var ConditionParser */
 	private $conditionParser;
 
@@ -49,53 +52,12 @@ class DisjunctionOperatorFunction implements IArrayFunction, IQueryBuilderFuncti
 		?IDbalAggregator $aggregator = null
 	): DbalExpressionResult
 	{
-		$isHavingClause = false;
-		$processedArgs = [];
-		$joins = [];
-
-		foreach ($this->normalizeFunctions($args) as $collectionFunctionArgs) {
-			$expression = $helper->processFilterFunction($builder, $collectionFunctionArgs, $aggregator);
-			$expression = $expression->applyAggregator($builder);
-			$processedArgs[] = $expression->getExpansionArguments();
-			$joins = array_merge($joins, $expression->joins);
-			$isHavingClause = $isHavingClause || $expression->isHavingClause;
-		}
-
-		return new DbalExpressionResult(
+		return $this->processQueryBuilderExpressionWithModifier(
 			'%or',
-			[$processedArgs],
-			$joins,
-			null,
-			$isHavingClause,
-			null,
-			null
+			$helper,
+			$builder,
+			$args,
+			$aggregator
 		);
-	}
-
-
-	/**
-	 * Normalizes directly entered column => value expression to expression array.
-	 * @phpstan-param array<string, mixed>|list<mixed> $args
-	 * @phpstan-return list<mixed>
-	 */
-	protected function normalizeFunctions(array $args): array
-	{
-		// Args passed as array values
-		// [ICollection::AND, ['id' => 1], ['name' => John]]
-		if (isset($args[0])) {
-			/** @phpstan-var list<mixed> $args */
-			return $args;
-		}
-
-		// Args passed as keys
-		// [ICollection::AND, 'id' => 1, 'name!=' => John]
-		/** @phpstan-var array<string, mixed> $args */
-		$processedArgs = [];
-		foreach ($args as $argName => $argValue) {
-			$functionCall = $this->conditionParser->parsePropertyOperator($argName);
-			$functionCall[] = $argValue;
-			$processedArgs[] = $functionCall;
-		}
-		return $processedArgs;
 	}
 }
